+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+- Remove support for Puppet older than 4.6.
+
 ## [1.2.0] - 2017-09-27
 - Support Conjur v5 API.